Add files via upload
In order for this file to work you would need to have a vcf file file already created. To be able to add, search, and print all functions to work, create a vcf file on your own. Once that vcf file is created, then you be to do the tasks.